Sun, Jul 3, 2022, 1:18 PM +00:00

+++ readability/readability.c
@@ -26,9 +26,9 @@ int main(void)
     printf("%i sentences\n", number_sentences);
 
     // Calculate the Coleman-Liau index
-    float L = (((float)number_letters / number_words) * 100);
-    float S = (((float)number_sentences / number_words) * 100);
-    float premature_index = (0.0588 * L) - (0.0296 * S) - 15.8;
+    float L = (number_letters / number_words) * 100;
+    float S = (number_sentences / number_words) * 100;
+    float premature_index = 0.0588 * L - 0.0296 * S -15.8;
     double index = round(premature_index);
 
     if (index < 1)
